backend/voice-detector/requirements.txt
#cuda==10.1
numpy==1.17.0
SoundFile
librosa==0.7.2
pyyaml==5.4
tensorboardX==2.1
matplotlib==3.0.2
